Make more examples into proper tests.
Note: `prepared` is combined into `basic`.
